refactor(redistribution): split into modules and name witnesses by role
Break the single 798-line redistribution.rs into cohesive modules and give
the witness slots names that express their function, removing the section
banner comments that signalled the file had outgrown its bounds.

- Split into anchor, neighbourhood, sample, witness and proof modules, each
  with its own tests; lib.rs holds the crate docs, SAMPLE_SIZE and the
  public re-exports.
- Rename the WitnessIndices fields require1/require2/require3 to
  `challenged: [usize; 2]` (the two slots the claim anchor pseudo-randomly
  selects to challenge) and `last` (the final slot, whose maximum transformed
  address anchors the reserve-size estimate). ChunkInclusionProofs now exposes
  `challenged()` and `last()` accessors in place of require1/2/3.

No behaviour or fixture bytes change; conformance, unit tests, clippy under
-D warnings, rustdoc and fmt all pass.
